requirement.txt
appdirs==1.4.4
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.4.2
blinker==1.4
Brlapi==0.6.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cronus==0.2
crypto==1.4.1
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
dnspython==2.0.0
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Mail==0.9.1
Flask-WTF==0.14.3
form==0.0.1
gunicorn==20.0.4
httplib2==0.9.2
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
Naked==0.1.31
netifaces==0.10.4
numpy==1.13.3
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
pipenv==2020.8.13
protobuf==3.0.0
PyAutoGUI==0.9.50
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGetWindow==0.0.8
pygobject==3.26.1
pylint==2.5.3
pymacaroons==0.13.0
PyMsgBox==1.0.8
PyMySQL==0.10.1
PyNaCl==1.1.2
pyperclip==1.8.0
PyRect==0.1.4
pyRFC3339==1.0
PyScreeze==0.1.26
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
python3-xlib==0.15
PyTweening==1.0.3
pytz==2018.3
pyxdg==0.25
PyYAML==5.3.1
reportlab==3.4.0
requests==2.24.0
requests-unixsocket==0.1.5
schedule==0.6.0
SecretStorage==2.3.1
shellescape==3.8.1
simple-crypt==4.1.7
simplejson==3.13.2
six==1.15.0
system-service==0.3
systemd-python==234
toml==0.10.1
typed-ast==1.4.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.3
validate-email==1.3
virtualenv==20.0.31
virtualenv-clone==0.5.4
wadllib==1.3.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
